feat: useBanner 훅 만들어서 로직 분리
