+++ src/components/templates/main/Banner.tsx
@@ -1,10 +1,10 @@
 'use client';
 
 import InnerLayout from '@/components/atoms/inner-layout/InnerLayout';
+import { useBanner } from '@/store/useBanner';
 import { Activity } from '@/types/activity';
 import Image from 'next/image';
 import Link from 'next/link';
-import { useEffect, useRef, useState } from 'react';
 
 const CIRCLE_LIST = [1, 2, 3];
 
@@ -13,51 +13,7 @@ interface Props {
 }
 
 export default function Banner({ activities }: Props) {
-  const [currentIndex, setCurrentIndex] = useState(1);
-  const slideRef = useRef<HTMLDivElement>(null);
-
-  const BG_NUM = activities.length;
-  const beforeSlide = activities[BG_NUM - 1];
-  const afterSlide = activities[0];
-
-  let slideArr = [beforeSlide, ...activities, afterSlide];
-  const SLIDE_NUM = slideArr.length;
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setCurrentIndex((prevIndex) => prevIndex + 1);
-    }, 8000);
-
-    return () => clearInterval(interval);
-  }, []);
-
-  useEffect(() => {
-    if (currentIndex === SLIDE_NUM - 1) {
-      setTimeout(() => {
-        if (slideRef.current) {
-          slideRef.current.style.transition = 'none';
-          setCurrentIndex(1);
-        }
-        setTimeout(() => {
-          if (slideRef.current) {
-            slideRef.current.style.transition = 'all 1000ms ease-in-out';
-          }
-        }, 100);
-      }, 1000);
-    } else if (currentIndex === 0) {
-      setTimeout(() => {
-        if (slideRef.current) {
-          slideRef.current.style.transition = 'none';
-          setCurrentIndex(BG_NUM);
-        }
-        setTimeout(() => {
-          if (slideRef.current) {
-            slideRef.current.style.transition = 'all 1000ms ease-in-out';
-          }
-        }, 100);
-      }, 1000);
-    }
-  }, [currentIndex, SLIDE_NUM, BG_NUM]);
+  const { currentIndex, slideRef, slideArr, setCurrentIndex } = useBanner(activities);
 
   return (
     <div className="relative h-240pxr w-full overflow-hidden bg-black md:h-550pxr">
